+++ src/Pim/Bundle/BaseConnectorBundle/Reader/File/CsvProductReader.php
@@ -7,7 +7,7 @@ use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\CurrencyRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
-use Pim\Bundle\TransformBundle\Builder\FieldNameBuilder;
+use Pim\Component\Connector\ArrayConverter\Flat\ProductAttributeFieldExtractor;
 
 /**
  * Product csv reader
@@ -26,8 +26,8 @@ class CsvProductReader extends CsvReader
     /** @var string[] Media attribute codes */
     protected $mediaAttributes;
 
-    /** @var FieldNameBuilder */
-    protected $fieldNameBuilder;
+    /** @var ProductAttributeFieldExtractor */
+    protected $fieldExtractor;
 
     /** @var ChannelRepositoryInterface */
     protected $channelRepository;
@@ -44,22 +44,22 @@ class CsvProductReader extends CsvReader
     /**
      * Constructor
      *
-     * @param EntityManager    $entityManager
-     * @param FieldNameBuilder $fieldNameBuilder
-     * @param string           $attributeClass
-     * @param string           $channelClass
-     * @param string           $localeClass
-     * @param string           $currencyClass
+     * @param EntityManager                  $entityManager
+     * @param ProductAttributeFieldExtractor $fieldExtractor
+     * @param string                         $attributeClass
+     * @param string                         $channelClass
+     * @param string                         $localeClass
+     * @param string                         $currencyClass
      */
     public function __construct(
         EntityManager $entityManager,
-        FieldNameBuilder $fieldNameBuilder,
+        ProductAttributeFieldExtractor $fieldExtractor,
         $attributeClass,
         $channelClass,
         $localeClass,
         $currencyClass
     ) {
-        $this->fieldNameBuilder    = $fieldNameBuilder;
+        $this->fieldExtractor    = $fieldExtractor;
         $this->attributeRepository = $entityManager->getRepository($attributeClass);
         $this->channelRepository   = $entityManager->getRepository($channelClass);
         $this->localeRepository    = $entityManager->getRepository($localeClass);
@@ -165,7 +165,7 @@ class CsvProductReader extends CsvReader
         $currencies = $this->currencyRepository->getActivatedCurrencyCodes();
 
         foreach ($this->fieldNames as $fieldName) {
-            if (null !== $info = $this->fieldNameBuilder->extractAttributeFieldNameInfos($fieldName)) {
+            if (null !== $info = $this->fieldExtractor->extractAttributeFieldNameInfos($fieldName)) {
                 $locale = $info['locale_code'];
                 $channel = $info['scope_code'];
                 $currency = isset($info['price_currency']) ? $info['price_currency'] : null;