+++ src/Pim/Component/Connector/ArrayConverter/Flat/AttributeOptionToStandardConverter.php
@@ -3,8 +3,8 @@
 namespace Pim\Component\Connector\ArrayConverter\Flat;
 
 use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
-use Pim\Component\Connector\Exception\ArrayConversionException;
 use Pim\Component\Connector\ArrayConverter\StandardArrayConverterInterface;
+use Pim\Component\Connector\Exception\ArrayConversionException;
 
 /**
  * Convert flat format to standard format for attribute option